==> scripts/05_timeOfDay_robustness.qmd <==
---
title: "05_timeOfDay_robustness"
format: html
editor: visual
---

```{r}
#| label: initalize
#| include: false

library(tidyverse)
library(emmeans)
library(ggplot2)
library(ggh4x)
library(cowplot)
library(patchwork)

eval <- FALSE # set this to FALSE to run nothing
save_figures <- FALSE
```

# Determining high-traffic stations

```{r, eval=eval}
#| label: stations_to_remove

regions <- c('arr', 'dfw', 'elp', 'hgb', 'san')
percent_removed <- 10/100 # 10% removed
resave_figs <- TRUE # helps with runtime

analysis <- lapply(regions, function(region){

  # Read data
  df <- readRDS(paste0('../data/dfs_Daily/',region)) |> 
    group_by(Station, TrafficType) |> 
    summarise(ADT = mean(TrafficCount), .groups = 'drop')
  
  # Get number of stations and number to remove
  n_stations <- df |> pull(Station) |> unique() |> length()
  n_to_remove <- ceiling(percent_removed * n_stations)
  
  # Get removed stations
  removed <- df |> 
    nest(data = c(Station, ADT), .by = TrafficType) |> 
    rowwise() |> 
    mutate(Station = list(data |> 
                            arrange(desc(ADT)) |> 
                            pull(Station) |> 
                            head(n = n_to_remove))) |>
    select(-data) |>
    unnest(cols = c(Station)) |> 
    pull(Station) |> 
    unique()

  result_detailed <- data.frame('Region' = region,
                                'Station' = removed)
  
  result_detailed
}) |> bind_rows()

write.csv(analysis, 
          '../data/high_traffic_stations.csv',
          row.names = F)
```

# Analysis function

This is for the case where the top 10% of high-traffic stations are removed and the analysis in 02 is rerun.

```{r, eval = eval}
#| label: analysis

regions <- c('arr', 'dfw', 'elp', 'hgb', 'san')
options(contrasts = c("contr.sum", "contr.poly"))

removed_stations <- read.csv('../data/high_traffic_stations.csv')

analysis <- lapply(regions, function(region){

  # Read data
  df <- readRDS(paste0('../data/dfs_timeOfDay/',region)) |> 
    mutate(IsOAD = factor(IsOAD, levels = c(1,0))) # for contrasts
  
  # Remove stations
  removed <- removed_stations |> 
    filter(Region == region) |> 
    pull(Station)
  
  df <- df |> 
    filter(!(Station %in% removed)) |> 
    droplevels()
  
  # Model equation
  eqn_quasi <- TrafficCount ~ TrafficType + Station + TimeOfDay + # traffic info
    IsOAD + AQI_Cat + # aqi
    Weekday + Holiday +  Month + Year + # datetime
    MeanT + Precip + Sunshine_Hrs + RelativeH + WindSpeed + WMO + # weather
    IsOAD:TimeOfDay + IsOAD:TrafficType + #No OAD/AQI int
    IsOAD:Holiday + # OAD/Holiday int +
    TimeOfDay:TrafficType + TimeOfDay:AQI_Cat + TrafficType:AQI_Cat + # other ints
    offset(log(Hours)) # adding offset
  
  # Fit model
  model_quasi <- glm(eqn_quasi, family = quasipoisson, data = df) 
  
  # Compute emmeans
  at = list(TrafficType = c("Bic","Ped")) 
  em <- emmeans(model_quasi, 
                ~ IsOAD + TimeOfDay + TrafficType + AQI_Cat + 
                  IsOAD:TimeOfDay + IsOAD:TrafficType + 
                  IsOAD:Holiday + 
                  TimeOfDay:TrafficType + TimeOfDay:AQI_Cat + TrafficType:AQI_Cat, 
                at = at,
                nuisance = c("Station","Weekday", "Month", "Year", 
                             "MeanT", "Precip","Sunlight_Hrs", "RelHum", "WindSpeed", "WMO")) 
  
  # Get out the pairwise contrasts that are important
  pair_0 <- pairs(em, 
                  by = c("TrafficType", "TimeOfDay","AQI_Cat","Holiday"),
                  adjust = "bonferroni") |> 
    summary(infer = c(TRUE, TRUE), level = 0.95) |> as.data.frame()
  
  # Estimate coefficents
  sum <- summary.glm(model_quasi) |> coef()
  
  # Get all params in the model
  all_params <- names(coef(model_quasi)) 
  
  # Get out the interaction terms 
  pair_1 <- pair_0 |> 
    mutate(TimeTraffic = interaction(TimeOfDay, TrafficType, Holiday,
                                        sep = " - ")) |> 
    dplyr::select(TrafficType, TimeOfDay, Holiday, estimate, 
                  asymp.LCL, asymp.UCL, p.value, TimeTraffic) |> 
    unique() |> 
    arrange(TimeOfDay)
  
# Save results
  filePath <- paste0("../results/05_timeOfDay_robustness/",region,"_contrasts.csv")
  write.csv(pair_1, filePath, row.names = FALSE)
  
  filePath <- paste0("../results/05_timeOfDay_robustness/",region,"_model_summary.csv")
  write.csv(sum, filePath)
})
```
